SQL Internship Project Files of Diabetes Prediction, HR Data Analysis & Pharma Data.
A 30 days internship by Psyliq. 
During this internship, had the opportunity to gain invaluable hands-on
experience in various aspects of data analysis and interpretation. 
The responsibility was over the tasks such as data collection, analysis, visualization, and deriving
meaningful insights. 
A unique opportunity to learn and actively contribute to our data-driven initiatives.